requirements.txt
openai
whisper
torch
pyannote.audio
watchdog
pystray
plyer
pydub
